create separate styled files for each componrmt in the info_page_shared folder
