+++ src/pages/info_page_shared/ButtonContainer.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./ButtonContainer.sc";
 
 export default function ButtonContainer({ children }) {
   return <s.ButtonContainer oneButton>{children}</s.ButtonContainer>;
