+++ src/pages/info_page_shared/ButtonContainer.sc.js
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+const ButtonContainer = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: center;
+  gap: 1rem;
+
+  @media (max-width: 768px) {
+    width: 100%;
+    flex-direction: column;
+    align-items: stretch;
+    gap: 0.75rem;
+  }
+`;
+
+export { ButtonContainer };
