+++ src/pages/info_page_shared/Footer.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./Footer.sc";
 
 export default function Footer({ children }) {
   return <s.Footer>{children}</s.Footer>;
