+++ src/pages/info_page_shared/Heading.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./Heading.sc";
 
 export default function Heading({ children }) {
   return <s.Heading>{children}</s.Heading>;
