+++ src/pages/info_page_shared/Heading.sc.js
@@ -1,0 +1,20 @@
+import styled from "styled-components";
+
+const Heading = styled.h1`
+  width: 80%;
+  font-size: 1.3em;
+  line-height: 150%;
+  text-align: center;
+  font-weight: 700;
+  margin: 0;
+  @media (max-width: 768px) {
+    width: 100%;
+    text-align: left;
+  }
+
+  @media (max-width: 576px) {
+    font-size: 1.2em;
+  }
+`;
+
+export { Heading };
