+++ src/pages/info_page_shared/Logo.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./Logo.sc";
 
 export default function Logo() {
   return <s.Logo src="../static/images/zeeguuLogo.svg" alt=""></s.Logo>;
