+++ src/pages/info_page_shared/Logo.sc.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Logo = styled.img`
+  width: 2.25rem;
+`;
+
+export { Logo };
