+++ src/pages/info_page_shared/Main.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./Main.sc";
 
 export default function Main({ children }) {
   return <s.Main>{children}</s.Main>;
