+++ src/pages/info_page_shared/Main.sc.js
@@ -1,0 +1,26 @@
+import styled, { css } from "styled-components";
+
+const BaseSectionStyle = css`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin: 1rem 0;
+  gap: 1.5rem;
+`;
+
+const Main = styled.main`
+  ${BaseSectionStyle}
+  line-height: 150%;
+
+  p {
+    width: 80%;
+  }
+
+  @media (max-width: 768px) {
+    p {
+      width: 100%;
+    }
+  }
+`;
+
+export { Main };
