+++ src/pages/info_page_shared/MainImage.js
@@ -1,4 +1,4 @@
-import * as s from "./InfoPage.sc";
+import * as s from "./MainImage.sc";
 
 export default function MainImage({ src }) {
   return <s.MainImage src={src} alt="" />;
