+++ src/pages/info_page_shared/MainImage.sc.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const MainImage = styled.img`
+  width: 80%;
+
+  @media (max-width: 768px) {
+    width: 100%;
+  }
+`;
+
+export { MainImage };
